Changes needed to make things compile.
Most changes are due to the implementation of proper void * pointers
over time which were not in Illumos-nexenta at the time of writing the
NDMP patches. Also zfs_nicenum moved to libzutil so we need to include
the header and link against that library. Some other small but obvious
fixes were needed for either the newer compiler used or pointer casting
needed.