Fix ndmpd_data_get_env_v4() call.
According to the RFC the ndmp_data_get_env call must return the
environment variables after both a backup and restore session. The old
implementation would only allow it after a backup.

This is what the NDMP RFC says:

This request is used by the DMA to obtain the backup environment
variable set associated with the current data operation. The
NDMP_DATA_GET_ENV request is typically issued following a successful
backup operation but MAY be issued during or after a recovery
operation as well. This request is only valid when the Data Server is
in the ACTIVE or HALTED states.

I also fixed some layout problems in this code file with extra empty
lines in switch statements and after a function. Things are now somewhat
more uniformly from a layout perspective.