Restore ZFS NDMP backups and restores
This reverts (parts of) the following Nexenta commits:

NEX-13374 NDMP should be able to backup unmounted ZFS filesystems
NEX-5801 Snapshots left over after failed backups

Lets for now remove these particular upstream patches from Nexenta for the
NDMPD as it removes the ZFS backup and restore method which some people
seem to be using. This reverts the needed bits but leaves most of the
cleanups in place.